+++ examples/nvexec/maxwell/std.cuh
@@ -23,7 +23,7 @@
 #include <barrier>
 
 template <class Shape>
-std::pair<Shape, Shape> even_share(Shape n, std::uint32_t rank, std::uint32_t size) noexcept {
+std::pair<Shape, Shape> even_share(Shape n, std::size_t rank, std::size_t size) noexcept {
   const auto avg_per_thread = n / size;
   const auto n_big_share = avg_per_thread + 1;
   const auto big_shares = n % size;
